pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "far-oracle"
version = "1.0.7"
description = "MCP Server for Federal Acquisition Regulations (FAR) search"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "blueskylineassets"}
]
keywords = ["mcp", "far", "federal-acquisition-regulations", "government-contracting", "ai-agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/blueskylineassets/far-mcp-server"
Repository = "https://github.com/blueskylineassets/far-mcp-server"
"RapidAPI" = "https://rapidapi.com/yschang/api/far-rag-federal-acquisition-regulation-search"

[project.scripts]
far-oracle = "server:main"

[tool.setuptools]
py-modules = ["server", "client"]
